+++ ios/Classes/AblyFlutter.m
@@ -599,6 +599,14 @@ static const FlutterHandler _restTime = ^void(AblyFlutter *const ably, FlutterMe
     }];
 };
 
+static const FlutterHandler _connectionRecoveryKey = ^void(AblyFlutter *const ably, FlutterMethodCall *const call, const FlutterResult result) {
+    AblyFlutterMessage *const ablyMessage = call.arguments;
+    AblyInstanceStore *const instanceStore = [ably instanceStore];
+    ARTRealtime *const realtime = [instanceStore realtimeFrom:ablyMessage.handle];
+    NSString *const connectionRecoveryKey = [realtime.connection createRecoveryKey];
+    result(connectionRecoveryKey);
+};
+
 static const FlutterHandler _getNextPage = ^void(AblyFlutter *const ably, FlutterMethodCall *const call, const FlutterResult result) {
     AblyFlutterMessage *const ablyMessage = call.arguments;
     
@@ -638,6 +646,30 @@ static const FlutterHandler _getFirstPage = ^void(AblyFlutter *const ably, Flutt
             NSNumber *const paginatedResultHandle = [instanceStore setPaginatedResult:paginatedResult handle:ablyMessage.handle];
             result([[AblyFlutterMessage alloc] initWithMessage:paginatedResult handle: paginatedResultHandle]);
         }
+    }];
+};
+
+static const FlutterHandler _realtimeAuthCreateTokenRequest = ^void(AblyFlutter *const ably, FlutterMethodCall *const call, const FlutterResult result) {
+    AblyFlutterMessage *const ablyMessage = call.arguments;
+    NSDictionary *const message = ablyMessage.message;
+    NSString *const channelName = (NSString*) message[TxTransportKeys_channelName];
+    ARTTokenParams *const tokenParams = (ARTTokenParams *) message[TxTransportKeys_params];
+    ARTAuthOptions *const authOptions = (ARTAuthOptions *) message[TxTransportKeys_options];
+    
+    AblyInstanceStore *const instanceStore = [ably instanceStore];
+    ARTRealtime *const realtime = [instanceStore realtimeFrom:ablyMessage.handle];
+    [realtime.auth requestToken:tokenParams withOptions:authOptions callback:^(ARTTokenDetails * _Nullable tokenDetails, NSError * _Nullable error) {
+        if (error) {
+            result([
+                    FlutterError
+                    errorWithCode:[NSString stringWithFormat: @"%ld", (long)error.code]
+                    message:[NSString stringWithFormat:@"Error creating token request = %@", error]
+                    details:error
+                    ]);
+        } else {
+            result(tokenDetails);
+        }
+        
     }];
 };
 
@@ -717,6 +749,8 @@ static const FlutterHandler _getFirstPage = ^void(AblyFlutter *const ably, Flutt
         AblyPlatformMethod_releaseRealtimeChannel: _releaseRealtimeChannel,
         AblyPlatformMethod_realtimeTime:_realtimeTime,
         AblyPlatformMethod_restTime:_restTime,
+        // Connection fields
+        AblyPlatformMethod_connectionRecoveryKey:_connectionRecoveryKey,
         // Push Notifications
         AblyPlatformMethod_pushActivate: PushHandlers.activate,
         AblyPlatformMethod_pushRequestPermission: PushHandlers.requestPermission,
@@ -732,6 +766,15 @@ static const FlutterHandler _getFirstPage = ^void(AblyFlutter *const ably, Flutt
         // Encryption
         AblyPlatformMethod_cryptoGetParams: CryptoHandlers.getParams,
         AblyPlatformMethod_cryptoGenerateRandomKey: CryptoHandlers.generateRandomKey,
+        //Authorize
+        AblyPlatformMethod_realtimeAuthAuthorize: AuthHandlers.realtimeAuthorize,
+        AblyPlatformMethod_realtimeAuthCreateTokenRequest: AuthHandlers.realtimeCreateTokenRequest,
+        AblyPlatformMethod_realtimeAuthRequestToken: AuthHandlers.realtimeRequestToken,
+        AblyPlatformMethod_realtimeAuthGetClientId: AuthHandlers.realtimeAuthClientId,
+        AblyPlatformMethod_restAuthAuthorize: AuthHandlers.restAuthorize,
+        AblyPlatformMethod_restAuthCreateTokenRequest: AuthHandlers.restCreateTokenRequest,
+        AblyPlatformMethod_restAuthRequestToken: AuthHandlers.restRequestToken,
+        AblyPlatformMethod_restAuthGetClientId: AuthHandlers.restAuthClientId
     };
 
     [registrar addApplicationDelegate:self];